@@ -2,12 +2,11 @@ import 'package:chitisplit/pages/add-expense.dart';
 import 'package:chitisplit/pages/add-person-to-group.dart';
 import 'package:chitisplit/pages/settings.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:chitisplit/classes/group.dart';
-import 'package:chitisplit/pages/add-expense.dart';
 
 class Home extends StatefulWidget {
+  const Home({Key? key}) : super(key: key);
   @override
   _HomeState createState() => _HomeState();
 }
@@ -17,16 +16,16 @@ class _HomeState extends State<Home> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Chi Ti Split'),
+        title: const Text('Chi Ti Split'),
         centerTitle: true,
         actions: <Widget>[
           IconButton(
-            icon: Icon(Icons.settings, color: Colors.white),
+            icon: const Icon(Icons.settings, color: Colors.white),
             onPressed: () {
               Navigator.push(
                 context,
                 MaterialPageRoute(
-                    builder: (context) => Settings(currentGroup: group)),
+                    builder: (context) => Settings(group)),
               ).then((value) => setState(() {}));
             },
           ),
@@ -40,9 +39,9 @@ class _HomeState extends State<Home> {
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
-              GroupOverview(currentGroup: group),
-              SizedBox(height: 20),
-              Menu(setParentState: this.setState, currentGroup: group),
+              GroupOverview(group),
+              const SizedBox(height: 20),
+              Menu(setState, group),
             ],
           ),
         ),
@@ -54,7 +53,7 @@ class _HomeState extends State<Home> {
 class GroupOverview extends StatefulWidget {
   final Group currentGroup;
 
-  GroupOverview({this.currentGroup});
+  const GroupOverview(this.currentGroup);
 
   @override
   _GroupOverviewState createState() => _GroupOverviewState();
@@ -65,7 +64,8 @@ class _GroupOverviewState extends State<GroupOverview> {
   Widget build(BuildContext context) {
     return Column(
       children: [
-        Text("Overview - ${group.currentUser}", style: TextStyle(fontSize: 25)),
+        Text("Overview - ${group.currentUser}",
+            style: const TextStyle(fontSize: 25)),
         DataTable(
           headingRowHeight: 15,
           columns: [
@@ -74,21 +74,21 @@ class _GroupOverviewState extends State<GroupOverview> {
           ],
           rows: [
             DataRow(cells: [
-              DataCell(Text("This event cost the group:")),
+              const DataCell(Text("This event cost the group:")),
               DataCell(Text("€${widget.currentGroup.totalGroupExpenses()}")),
             ]),
             DataRow(cells: [
-              DataCell(Text("It cost you:")),
+              const DataCell(Text("It cost you:")),
               DataCell(Text(
                   "€${widget.currentGroup.totalExpenses(widget.currentGroup.currentUser)}")),
             ]),
             DataRow(cells: [
-              DataCell(Text("You have paid:")),
+              const DataCell(Text("You have paid:")),
               DataCell(Text(
                   "€${widget.currentGroup.totalPayments(widget.currentGroup.currentUser)}")),
             ]),
             DataRow(cells: [
-              DataCell(Text("You owe:")),
+              const DataCell(Text("You owe:")),
               DataCell(Text(
                   "€${widget.currentGroup.personalBalance(widget.currentGroup.currentUser)}")),
             ]),
@@ -103,33 +103,32 @@ class Menu extends StatelessWidget {
   final void Function(void Function()) setParentState;
   final Group currentGroup;
 
-  Menu({this.setParentState, this.currentGroup});
+  const Menu(this.setParentState, this.currentGroup);
 
   @override
   Widget build(BuildContext context) {
     return Column(children: [
       Card(
         child: ListTile(
-          leading: Icon(MdiIcons.pencilPlus),
-          trailing: Icon(Icons.play_arrow),
-          title: Text('Add expense', textAlign: TextAlign.center),
+          leading: const Icon(MdiIcons.pencilPlus),
+          trailing: const Icon(Icons.play_arrow),
+          title: const Text('Add expense', textAlign: TextAlign.center),
           onTap: () {
             Navigator.push(
               context,
-              MaterialPageRoute(
-                  builder: (context) => AddExpense(currentGroup: currentGroup)),
+              MaterialPageRoute(builder: (context) => AddExpense(currentGroup)),
             ).then((value) => setParentState(() {}));
           },
         ),
       ),
-      Card(
+      const Card(
         child: ListTile(
           leading: Icon(MdiIcons.cashRefund),
           trailing: Icon(Icons.play_arrow),
           title: Text('Transfer', textAlign: TextAlign.center),
         ),
       ),
-      Card(
+      const Card(
         child: ListTile(
           leading: Icon(MdiIcons.formatListBulletedSquare),
           trailing: Icon(Icons.play_arrow),
@@ -138,16 +137,15 @@ class Menu extends StatelessWidget {
       ),
       Card(
         child: ListTile(
-          leading: Icon(Icons.person_add),
-          trailing: Icon(Icons.play_arrow),
-          title: Text('Add person to group', textAlign: TextAlign.center),
+          leading: const Icon(Icons.person_add),
+          trailing: const Icon(Icons.play_arrow),
+          title: const Text('Add person to group', textAlign: TextAlign.center),
           onTap: () {
             Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (context) =>
-                      AddPersonToGroup(currentGroup: currentGroup)),
-            ).then((value) => setParentState(() {}));
+                    context,
+                    MaterialPageRoute(
+                        builder: (context) => AddPersonToGroup(currentGroup)))
+                .then((value) => setParentState(() {}));
           },
         ),
       ),
